dom/tests/mochitest/chrome/test_fullscreen_preventdefault.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  Test for fullscreen sizemode in chrome
  -->
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        sizemode="fullscreen">

  <script type="application/javascript" 
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      

<script>
SimpleTest.waitForExplicitFinish();

newwindow = window.open("fullscreen_preventdefault.xul", "_blank","chrome,resizable=yes");

function done()
{
  // because we are cancelling the fullscreen event, it
  // takes a bit for the fullScreen property to be set
  setTimeout("complete()", 0);
}

function complete()
{
  ok(!(newwindow.fullScreen), "window.fullScreen is false.");
  newwindow.close();
  SimpleTest.finish();
}

</script>


<body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;"/>

</window>
